Move groupAccess to separate file and add test
